feat(app/collab): refactor auth plugin dir structure
