+++ apps/api/package.json
@@ -6,10 +6,11 @@
 	"scripts": {
 		"build": "npm run clean && tsc",
 		"clean": "rm -rf build",
-		"dev": "npm run build && nodemon --config nodemon.json",
+		"dev": "nodemon --config nodemon.json",
 		"lint": "biome lint ./src",
 		"start": "node build/src/index",
-		"test": "NODE_ENV=test vitest --run",
+		"test": "NODE_ENV=test vitest --run --coverage",
+		"test:coverage": "NODE_ENV=test vitest --run --coverage",
 		"test:watch": "NODE_ENV=test vitest",
 		"typecheck": "tsc --noEmit --incremental false"
 	},
@@ -34,6 +35,7 @@
 		"fastify-plugin": "^4.5.1",
 		"googleapis": "^133.0.0",
 		"open-graph-scraper": "^6.2.2",
+		"pino": "^8.20.0",
 		"pino-pretty": "^10.2.0",
 		"zod": "^3.21.4"
 	},
@@ -41,6 +43,7 @@
 		"@types/node": "^20.4.5",
 		"nodemon": "^3.0.1",
 		"ts-node": "10.4.0",
+		"tsconfig-paths": "^4.2.0",
 		"typescript": "5.4.3",
 		"vitest": "^1.4.0"
 	}
