+++ apps/api/src/index.ts
@@ -1,3 +1,6 @@
+import dotenv from "dotenv";
+dotenv.config();
+
 import "@total-typescript/ts-reset";
 import { startServer } from "./server";
