--- apps/api/src/plugins/auth/constants.ts
@@ -1,8 +1,0 @@
-export const TOKEN_FAILURE_REASONS = {
-	EXPIRED: "expired",
-	INVALID: "token_invalid",
-	NOT_FOUND: "token_not_found",
-	EMAIL_MISMATCH: "email_mismatch",
-} as const;
-
-export default {};
