+++ apps/api/src/plugins/auth/logout/index.ts
@@ -1,0 +1,10 @@
+import type { FastifyPluginAsync } from "fastify";
+
+const logoutPlugin: FastifyPluginAsync = async function logoutPlugin(server) {
+  server.post("/logout", {}, async (request, reply) => {
+    request.session.delete();
+    return reply.code(200).send();
+  });
+}
+
+export default logoutPlugin;
