+++ apps/api/src/plugins/invites/get/getUserInvites.test.ts
@@ -13,6 +13,10 @@ describe("GET /invites", () => {
 		server = await createServer();
 	});
 
+	afterAll(async () => {
+		await server.close();
+	});
+
 	afterEach(() => {
 		vi.clearAllMocks();
 	});
