+++ apps/api/src/plugins/lists/delete/index.ts
@@ -1,8 +1,7 @@
+import { prisma } from "@hominem/db";
 import type { FastifyInstance } from "fastify";
 
-import { prisma } from "@hominem/db";
-
-import { verifySession } from "../../auth";
+import { verifySession } from "../../auth/utils";
 
 const deleteListRoute = (server: FastifyInstance) => {
 	server.delete(
