+++ apps/api/src/plugins/lists/index.ts
@@ -25,14 +25,14 @@ const listsPlugin: FastifyPluginAsync = async (server) => {
 	putListRoute(server);
 
 	// Cron jobs
-	if (process.env.NODE_ENV !== "test") {
+	// if (process.env.NODE_ENV !== "test") {
 		// addPhotoToPlaces(server).catch((err) => {
 		//   console.error("Error adding photo to place", err);
 		// });
 		// migrateLatLngFloat(server).catch((err) => {
 		//   console.error("Error migrating lat and lng", err);
 		// });
-	}
+	// }
 };
 
 export default fastifyPlugin(listsPlugin);
