+++ apps/api/src/plugins/places/place.test.ts
@@ -1,13 +1,15 @@
 import { prisma } from "@hominem/db";
 import type { FastifyInstance } from "fastify";
+import type { Mock } from "vitest";
 import { vi } from "vitest";
 
 import { createServer } from "@app/server";
 import { MOCKS } from "@test/mocks";
-import { googlePlaces } from "@test/test.setup";
 import { mockAuthSession } from "@test/utils";
 
 import * as googlePlacesPlugin from "../google/places";
+
+const googlePlaces = googlePlacesPlugin.places;
 
 const { PLACE, GOOGLE_PLACE_GET, GOOGLE_PHOTO_MEDIA } = MOCKS;
 
@@ -34,8 +36,8 @@ describe("/places/:id", () => {
 
 		test("should return place from Google Places if place does not exist in DB", async () => {
 			vi.spyOn(prisma.place, "findFirst").mockResolvedValue(null);
-			googlePlaces.get.mockResolvedValue(GOOGLE_PLACE_GET);
-			googlePlaces.photos.getMedia.mockResolvedValue(GOOGLE_PHOTO_MEDIA);
+			(googlePlaces.get as Mock).mockResolvedValue(GOOGLE_PLACE_GET);
+			(googlePlaces.photos.getMedia as Mock).mockResolvedValue(GOOGLE_PHOTO_MEDIA);
 			const response = await server.inject({
 				method: "GET",
 				url: "/places/123",
